Monolith.Backend.Services.API

Description

This module contains the interface for services which provide the actual API functionality of the Monolith backend: running a web server and responding to requests for various data.

Documentation

newtype API

Handle for an API service instance.

Constructors

API

Fields

getAsync :: Async ()

The Async reference to the thread that started the API service.