Monolith.Backend.Services.API.Scotty

Description

This module implements the Monolith.Backend.Services.API interface with Scotty.

Documentation

getHandle :: RealtimeData -> Port -> IO API

Create a new API. This kicks off a new thread and starts up a Scotty in it. Presumably events coming into this web server will drive the rest of this application; all services should be written with concurrency in mind.