Monolith.Backend.Services.API.Utilities

Description

This module contains things that may be useful for any service that implements the Monolith.Backend.Services.API interface, like (for instance) the function getTickerText which turns Routes into text for the ticker.

Documentation

getTickerText :: Int -> [Route] -> String