Monolith.Backend.Services.RealtimeData

Description

This is the interface for modules that provide realtime transit service information to other services in the Monolith backend. This service is specific to realtime data; static data should come from another service.

Documentation

newtype RealtimeData

Handle for a real-time data service.

Constructors

RealtimeData

Fields

incomingTripsForStop :: StopID -> IO Stop

Get a Stop object with real-time trip data for the given stop ID.

data RealtimeDataException

General-purpose exception to throw when something goes wrong fetching real-time data.

Constructors

RealtimeDataException !String

type StopID = Text