Monolith.Backend.Services.RealtimeData.Cache

Description

This service implements the RealtimeData interface. It sits on another implementation of the same interface, and acts as a caching layer that allows previously computed results within some staleness threshold to be re-used.

Documentation

data CacheConfig

Configuration type for this service

Constructors

CacheConfig

Fields

expirationTime :: !Int

The expiration time of cache data, in seconds

defaultCacheConfig :: CacheConfig

newRealtimeCache :: CacheConfig -> RealtimeData -> IO RealtimeData

Create a new Cache service wrapping another realtime data service.