Monolith.Backend.Services.RealtimeData.ObaRest.HTTP

Documentation

jsonForRouteAndParams :: FromJSON j => ObaRestConfig -> String -> String -> [(String, String)] -> IO (Either String j)

The OBA REST API methods are basically foo.barbazMETHOD/ID.json?params=things This function takes care of putting that behind a reasonably usable interface, and will result in Either an error String or an Aeson FromJSON instance.