Monolith.Backend.Services.RealtimeData.Types

Description

This module contains the types and instances that pertain directly to the data served up by the REST API. They do not apply outside of that context.

Documentation

data Stop

This type encapsulates some information about a stop, and contains a list of routes and upcoming trips for those routes. (and so on). Since this stop data is tied to a certain time, the stopTimestamp member records the time when this datum was generated.

We use Sets here so that there's no possibility for things to get out of order. The correct ordering is important.

Constructors

Stop

Fields

stopId :: Text

stopDesc :: Text

stopRoutes :: [Route]

stopTimestamp :: Int

data Route

Constructors

Route

Fields

routeId :: Text

routeNumber :: Text

routeDesc :: Text

routeTrips :: Set Trip

data Trip

Constructors

Trip

Fields

tripArrival :: Int

tripId :: Text

tripRouteId :: Text

tripWaitSource :: WaitSource

tripHeadSign :: Text

data WaitSource

Constructors

Realtime

Scheduled