Monolith.Backend.Services.StaticData

Description

This is the interface for modules that provide static transit service information to other services in the Monolith backend. Note that services implementing this interface can cheat, and use whatever data sources they want...

As we develop it, this interface should basically be a layer on top of GTFS (static) or equivalent; it should provide static data concerning the transit infrastructure.

Documentation

data StaticData

Constructors

StaticData

Fields

getStopForId :: StopId -> IO Stop

stopsWithinRadius :: Point -> Double -> IO [Stop]

stopsWithinRadiusOfStop :: StopId -> Double -> IO StopVicinity

module Monolith.Backend.Services.StaticData.Types