Monolith.Backend.Services.StaticData.ObaRest

Documentation

newHandle :: ObaRestConfig -> IO StaticData

getStopForId' :: ObaRestConfig -> IO (StopId -> IO Stop)

stopsWithinRadius' :: ObaRestConfig -> IO (Point -> Double -> IO [Stop])

We're building in a perma-cache here because this is in theory static data that will only change once a month or even less often. This whole module is a hack that will go away when we get GTFS data integrated, so don't worry too much about the implications of this. In any case, it shouldn't take up too much memory in normal operation.

stopsWithinRadiusOfStop' :: ObaRestConfig -> IO (StopId -> Double -> IO StopVicinity)

This function searches for a stop by ID and then returns information about what's in the vicinity. It's intended to be used to populate a map view nearby the kiosk or similar.

stopsWithinRadiusOfStop'' :: ObaRestConfig -> (StopId, Double) -> IO StopVicinity