Monolith.Backend.Services.StaticData.ObaRest.Utilities

Documentation

getStops :: ObaRestConfig -> (Double, Double, Double) -> IO [Stop]

Get a list of stops from the OBA API for the given lon, lat, and radius.

getStop :: ObaRestConfig -> StopId -> IO Stop

Get a Stop for the given ID from the OBA API.

makeStop :: Value -> Value -> Maybe Stop

with a "stop blob" Value and a "routes blob" Value, make a new Stop.

fillStopRoutes :: Stop -> Value -> Maybe Stop

For a given stop and a Value containing a list of OBA route "references", replace route ids in the stop's routes attribute with the route names that correspond to those IDs.

addDistanceToHomeStop :: Stop -> Stop -> Maybe Stop

This function is used to compute the distance between two Stops, and to annotate one of the Stop arguments with the result.