Monolith.Backend.Services.StaticData.Types

Description

This module contains outward-facing types relevant to the static data service interface, which is basically responsible for providing info like which stops are where, which routes service which stops, and so on.

Exceptions

data StaticDataException

Constructors

StaticDataException !String

Point data type and lenses

data Point

Constructors

Point

Fields

_pointLon :: !Double

_pointLat :: !Double

pointLon :: Lens' Point Double

pointLat :: Lens' Point Double

Stop data type and lenses

type Route = Text

type StopId = Text

data Stop

Constructors

Stop

Fields

_stopId :: !StopId

_stopName :: !Text

_stopLocation :: !Point

_stopDirection :: !Text

_stopRoutes :: ![Route]

_stopDistanceFromHomeStop :: !(Maybe Double)

stopId :: Lens' Stop StopId

stopName :: Lens' Stop Text

stopLocation :: Lens' Stop Point

stopDirection :: Lens' Stop Text

stopRoutes :: Lens' Stop [Route]

stopDistanceFromHomeStop :: Lens' Stop (Maybe Double)

Stop vicinity data type and lenses

data StopVicinity

Constructors

StopVicinity

Fields

_vicinityHomeStop :: Stop

_vicinityRadius :: Double

_vicinityNearbyStops :: [Stop]

_vicinityBikeShares :: [()]

_vicinityCarShares :: [()]

_vicinityEvents :: [()]

vicinityHomeStop :: Lens' StopVicinity Stop

vicinityRadius :: Lens' StopVicinity Double

vicinityNearbyStops :: Lens' StopVicinity [Stop]

vicinityBikeShares :: Lens' StopVicinity [()]

vicinityCarShares :: Lens' StopVicinity [()]

vicinityEvents :: Lens' StopVicinity [()]