Monolith.Utility.Cache

Description

This module contains the Cache type, which is basically a wrapper around a read/write lock and some kind of map. These Caches are oriented toward caching the results of expensive IO actions.

Documentation

class CacheMap m k v | m -> k, m -> v

Typeclass for map structures that can be used in a Cache

Minimal complete definition

empty, mapLookup, mapInsert

data Cache m k v

type HashCache k v = Cache (HashMap k v) k v

newCache :: CacheMap m k v => (k -> IO v) -> IO (Cache m k v)

cacheLookup :: CacheMap m k v => k -> Cache m k v -> IO v

Lookup a key in a Cache, running the default function and caching its result if the key is not known.