Monolith.Backend.Services.RealtimeData.Types

Description

This module contains the types and instances that pertain directly to the data served up by the REST API. They do not apply outside of that context.

Documentation

data Stop

This type encapsulates some information about a stop, and contains a list of routes and upcoming trips for those routes. (and so on). Since this stop data is tied to a certain time, the stopTimestamp member records the time when this datum was generated.

We use Sets here so that there's no possibility for things to get out of order. The correct ordering is important.

Constructors

Stop !Text !Text ![Route] !Int

stopId :: Lens' Stop Text

stopDesc :: Lens' Stop Text

stopRoutes :: Lens' Stop [Route]

stopTimestamp :: Lens' Stop Int

data Route

Constructors

Route !Text !(Maybe Int) !Text !Text !(Set Trip)

routeId :: Lens' Route Text

earliestTrip :: Lens' Route (Maybe Int)

routeNumber :: Lens' Route Text

routeDesc :: Lens' Route Text

routeTrips :: Lens' Route (Set Trip)

data TripWaitTime

Constructors

TripDue Int

TripArrivesIn Int

data Trip

Constructors

Trip !Int !(Maybe TripWaitTime) !Text !Text !WaitSource !Text

tripArrival :: Lens' Trip Int

tripWaitTime :: Lens' Trip (Maybe TripWaitTime)

tripId :: Lens' Trip Text

tripRouteId :: Lens' Trip Text

tripWaitSource :: Lens' Trip WaitSource

tripHeadSign :: Lens' Trip Text

data WaitSource

Constructors

Realtime

Scheduled